Split template module into multiple files.
